requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
kombu==5.5.4
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==5.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
